requirements.txt
pandas
matplotlib
prophet
requests
python-dotenv
